pyproject.toml
[project]
name = "falkordb-fastmcp"
version = "1.0.0"
description = "FastMCP-based Model Context Protocol server for FalkorDB graph database"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.3.0",
    "falkordb>=1.2.0",
    "python-dotenv>=1.0.0",
    "claude-agent-sdk>=0.1.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/falkordb_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
